requirements.txt
gunicorn==20.0.4
Flask==1.1.2
Jinja2==2.11.2
python-dotenv==0.14.0
pytest==6.0.0
requests==2.24.0